Add movie actions to their own file and adjust imports accordingly
