+++ src/containers/Login/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import './Login.css';
+import { getFavorites } from '../../actions/movieActions';
 import * as userActions from '../../actions/userActions';
 import './Login.css';
 import * as API from '../../utilities/API';
@@ -40,7 +41,7 @@ export class Login extends Component {
 
   getFavoritesFromDatabase = async (user) => {
     const retrievedFavorites = await API.getFavorites(user)
-    userActions.getFavorites(retrievedFavorites)
+    getFavorites(retrievedFavorites)
   }
 
   toggleCreate = async (event) => {
@@ -49,7 +50,7 @@ export class Login extends Component {
     if (create) {
       const message = await API.createUser({email, password, name});
       this.setState({error: 'Email has already been used'})
-    } 
+    }
     this.setState({ create: true});
   }
 
@@ -80,15 +81,15 @@ export class Login extends Component {
           type='text'
           placeholder='Password'
           onChange={this.handleChange}/>
-        { !create 
-            ? <button 
+        { !create
+            ? <button
                 type='submit'
                 className='login-button'
                 onClick={this.submitLogin}>Login
               </button>
             : null }
-        <button 
-          className='create-user' 
+        <button
+          className='create-user'
           onClick={this.toggleCreate}>Create User
         </button>
         { error.length ? <div className="error">{error}</div> : null }
